sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=66", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soulbrowser-sdk"
version = "0.1.0"
description = "Preview Python SDK for the SoulBrowser testing console APIs"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "SoulBrowser" }
]
dependencies = [
    "httpx>=0.27",
]

[project.optional-dependencies]
ws = ["websocket-client>=1.7"]

[project.urls]
homepage = "https://github.com/SoulBrowserClaude/SoulBrowser"
